Dealcloud.Management (dealcloud v0.2.0)

Management contains all of the apis needed to work with users

With these APIs you are able to:

Get Users, ~> By email, By modified data, by Id

Create Users

Update users ~> change users names, active/inactive, change email addresses, change user permissions

Delete users

Get User groups

Get user group details

Get user activities

Get proxy assignments

Create proxy assignments

Types

activityQuery()

Specs

activityQuery() :: %Dealcloud.Management.ActivityQuery{
  activity: term(),
  dateFrom: term(),
  dateTo: term(),
  source: term(),
  userIds: term()
}

license()

Specs

license() :: %Dealcloud.Management.License{
  description: term(),
  licenseType: term(),
  userId: term()
}

proxy()

Specs

proxy() :: %Dealcloud.Management.Proxy{proxyUserId: term(), userIds: term()}

proxyQuery()

Specs

proxyQuery() :: %Dealcloud.Management.ProxyQuery{
  pageNumber: term(),
  pageSize: term(),
  proxyQuery: term(),
  searchByName: term(),
  userQuery: term()
}

query()

Specs

query() :: %Dealcloud.Management.Query{
  email: term(),
  isBillable: term(),
  modifiedSince: term(),
  pageNumber: term(),
  pageSize: term()
}

user()

Specs

user() :: %Dealcloud.Management.User{
  createdBy: term(),
  createdDate: term(),
  email: term(),
  federationId: term(),
  firstName: term(),
  group: term(),
  id: term(),
  isActive: term(),
  isLocked: term(),
  lastName: term(),
  modifiedBy: term(),
  modifiedDate: term()
}

Functions

create_user_proxy(body, config)

Specs

create_user_proxy(Dealcloud.Management.Proxy, Dealcloud.Auth) :: any()

This creates new proxy assignments. There's no way to update an existing proxy assignments. Proxy assignments can be replaced

Examples

Dealcloud.Management.get_user_proxy(%Dealcloud.Management.Proxy{proxyUserId: 1, userIds: [2,3,4]}, Dealcloud.Auth)

create_users(body, config)

Specs

create_users([User], Dealcloud.Auth) :: any()

Takes a list of Dealcloud.Management.user() and creates them as new

Examples

Dealcloud.Management.create_users([%User{ firstName: "James", lastName: "King", email: "kingj@basketball.com"}], Dealcloud.Auth)

delete_user(userId, config)

Specs

delete_user(binary(), Dealcloud.Auth) :: any()

Takes a user id and deletes the associated user

Examples

Dealcloud.Management.delete_users(1, Dealcloud.Auth)

get_by_ids(param, config)

Specs

get_by_ids([integer()], Dealcloud.Auth) :: any()

Takes a list of integers representing the user's Id and returns the users

Examples

Dealcloud.Management.get_by_ids([1,2,3], Dealcloud.Auth)

get_user(param, config)

Specs

get_user(Dealcloud.Management.Query, Dealcloud.Auth) :: any()

Takes a query containing either the user's email address or a dateTime, attempts to find the user or return all user modified since the dateTime provided

Examples

iex> Dealcloud.Management.get_user(%Dealcloud.Management.Query{email: "test@email.com"}, Dealcloud.Auth)

or

iex> Dealcloud.Management.get_user(%Dealcloud.Management.Query{ModifiedSince: "2019-07-15T15:53:00.050Z"}, Dealcloud.Auth)

get_user_activity(body, config)

Specs

get_user_activity(Dealcloud.Management.ActivityQuery, Dealcloud.Auth) :: any()

These are the user activity APIs ## Examples

Dealcloud.Management.get_user_activity(%Dealcloud.Management.ActivityQuery{},%Dealcloud.Auth{})

get_user_activity(body, query, config)

Specs

get_user_activity(
  Dealcloud.Management.ActivityQuery,
  Dealcloud.Management.UserQuery,
  Dealcloud.Auth
) :: any()

These are the user activity APIs ## Examples

Dealcloud.Management.get_user_activity(%Dealcloud.Management.ActivityQuery{},%Dealcloud.Management.UserQuery{}, %Dealcloud.Auth{})

get_user_group(groupId, config)

Specs

get_user_group(integer(), Dealcloud.Auth) :: any()

These are the user group APIs

Examples

Dealcloud.Management.get_user_group(1, Dealcloud.Auth)

get_user_groups(config)

Specs

get_user_groups(Dealcloud.Auth) :: any()

These are the user group APIs

Examples

Dealcloud.Management.get_user_groups(Dealcloud.Auth)

get_user_license(userId, config)

Specs

get_user_license(integer(), Dealcloud.Auth) :: any()

These are the user license APIs

Examples

Dealcloud.Management.get_user_license(1, Dealcloud.Auth)

get_user_proxy(query, config)

Specs

get_user_proxy(Dealcloud.Management.ProxyQuery, Dealcloud.Auth) :: any()

These are the user proxy APIs

There are two ways to query for proxies, you can search by user or by the proxy

Examples

Dealcloud.Management.get_user_proxy(%Dealcloud.Management.ProxyQuery{proxyQuery: "test@user.com"}, Dealcloud.Auth)
Dealcloud.Management.get_user_proxy(%Dealcloud.Management.ProxyQuery{userQuery: "test@user.com"}, Dealcloud.Auth)

get_users_license(config)

Specs

get_users_license(Dealcloud.Auth) :: any()

Get all licenses

Examples

Dealcloud.Management.get_user_license(Dealcloud.Auth)

get_users_license(params, config)

Specs

get_users_license(Dealcloud.Management.Query, Dealcloud.Auth) :: any()

Get licenses for multiple users

Examples

Dealcloud.Management.get_users_license(%Dealcloud.Management.Query{isBillable: true}, Dealcloud.Auth)

update_user(body, config)

Specs

update_user(integer(), Dealcloud.Auth) :: any()

Takes a list of Dealcloud.Management.user() and updates them

Examples

Dealcloud.Management.update_users([%User{ id: 1, firstName: "James", lastName: "King", email: "kingj@basketball.com"}], Dealcloud.Auth)

update_user_license(license, config)

Specs

update_user_license(Dealcloud.Management.License, Dealcloud.Auth) :: :ok

Update licenses

Examples

Dealcloud.Management.update_user_license(%Dealcloud.Management.License{userId: 1, licenseType: 1}, Dealcloud.Auth)

user_get_by_id(param, config)

Specs

user_get_by_id(integer(), Dealcloud.Auth) :: any()

Takes a integer representing the user's Id and returns the user

Examples

iex> Dealcloud.Management.user_get_by_id(1, Dealcloud.Auth)