Dealcloud.Schema (dealcloud v0.2.0)

Schema tells you information about your site and about the system.

Use these APIs to discover things you need to know about configurations that are possible as well as what's currently implemented

Functions

create_choices(fieldId, choices, config)

Specs

create_choices(integer(), [binary()], Dealcloud.Auth.t()) :: any()

Appends new choice values to choice fields

Schema.create_choices(["Active", "Inactive"], %Auth{})

get_all_fields(config)

Specs

get_all_fields(Dealcloud.Auth.t()) :: any()

Returns all fields from the site. Use this sparingly, this is heavy operation only needed when generating full site schema.

Schema.get_entry_fields("company", %Auth{}) Schema.get_entry_fields(2011, %Auth{})

get_currencies(config)

Specs

get_currencies(Dealcloud.Auth.t()) :: any()

These are the system Information APIs

Returns all currencies supported by the system Schema.currencies(%Auth{})

get_entry_fields(entryType, config)

Specs

get_entry_fields(binary() | integer(), Dealcloud.Auth.t()) :: any()

Returns fields for a specific entry type

Schema.get_entry_fields("company", %Auth{}) Schema.get_entry_fields(2011, %Auth{})

get_entry_type(entryType, config)

Specs

get_entry_type(binary() | integer(), Dealcloud.Auth.t()) :: any()

Returns information about specific entry type

Schema.get_entry_types("company", %Auth{}) Schema.get_entry_types(2011, %Auth{})

get_entry_types(config)

Specs

get_entry_types(Dealcloud.Auth.t()) :: any()

These are the site Information APIs

Returns information about all entry types configured in the site

Schema.get_entry_types(%Auth{})

get_field(fieldId, config)

Specs

get_field(integer(), Dealcloud.Auth) :: any()

Returns information of a field. Requires the numerical ID of the field

Schema.get_field(3, %Auth{})

get_field_types(config)

Specs

get_field_types(Dealcloud.Auth.t()) :: any()

Returns information about field types. This information is helpful when looking at the entry type fields. Fields configured can be any of the fields returned by the endpoint

Schema.get_system_entry_type(%Auth{})

get_fields(fields, config)

Specs

get_fields([integer()], Dealcloud.Auth) :: any()

Returns information of a selection of fields passed in as query params

Schema.get_fields([3,4,5], %Auth{})

get_filters(config)

Specs

get_filters(Dealcloud.Auth.t()) :: any()

Returns information about all the filters supported by the data apis

Schema.get_filters(%Auth{})

get_system_entry_type(config)

Specs

get_system_entry_type(Dealcloud.Auth.t()) :: any()

Returns information about all the system entry types

Schema.get_system_entry_type(%Auth{})

get_system_field_type(config)

Specs

get_system_field_type(Dealcloud.Auth.t()) :: any()

Returns information about all the system field types

Schema.get_system_field_type(%Auth{})

get_time_zone(config)

Specs

get_time_zone(Dealcloud.Auth.t()) :: any()

Returns information about all timezones supported by the system

Schema.get_time_zone(%Auth{})

get_users(config)

Specs

get_users(Dealcloud.Auth.t()) :: any()

Returns all users in the site.

Schema.get_field(%Auth{})

get_users(params, config)

Specs

get_users(map(), Dealcloud.Auth.t()) :: any()

Returns all users in the site filtered by ativeOnly filter

Schema.get_field(%{"activeOnly" => "true"}, %Auth{})