Dealcloud.Data.Cells (dealcloud v0.2.0)

These are cell APIs. These APIs use key-value pairs when communicating with the server. Both Requests and Responses are sent as arrays of [{EntryId: 1, FieldId: 1, Value: "Hello"}].

These APIs are more performant, but may not be the best when working with ETL tools.

Functions

all_fields(entryType, entryIds, config)

See Dealcloud.Impl.Data.Utility.all_fields/3.

create(type, entries, config)

See Dealcloud.Impl.Data.EntryData.post/3.

get(entries, params, config)

See Dealcloud.Impl.Data.EntryData.get/3.

get_all(entryType, config)

See Dealcloud.Impl.Data.Utility.all_entries/2.

get_batch(entries, params, config)

See Dealcloud.Impl.Data.Utility.get/3.

get_filtered(entryType, filters, config)

See Dealcloud.Impl.Data.EntryData.filter_entries/3.

get_ids(entryType, config)

See Dealcloud.Impl.Data.EntryData.get_entries_ids/2.

get_keys(entryType, config)

See Dealcloud.Impl.Data.EntryData.get_entries/2.

get_modified(entryType, params, config)

See Dealcloud.Impl.Data.EntryData.get_modified_entries/3.

update(type, entries, config)

See Dealcloud.Impl.Data.EntryData.put/3.